requirements.txt
customtkinter==5.2.2
pillow=12.0.0